replica/m365/apps/requirements.txt
flask
pyjwt==2.8.0
requests
gunicorn
# For production, add:
# msal
